requirement.txt
streamlit
 pandas
sklearn.feature_extraction.text import TfidfVectorizer
sklearn.metrics.pairwise import cosine_similarity
 nltk
nltk.corpus import stopwords
 string
  base64